倒数第二次

+++ src/sunbookstore/order/service/DeleteOrderService.java
@@ -5,11 +5,11 @@ import sunbookstore.order.dao.OrderDao;
 public class DeleteOrderService {
 	OrderDao orderDao = new OrderDao();
 	//删除订单的明细
-	public int deleteBook_Order(int oid) {
+	public int deleteBook_Order(double oid) {
 		return orderDao.deleteBook_Order(oid);
 	}
 	//删除订单
-	public int deleteOrder(int oid) {
+	public int deleteOrder(double oid) {
 		return orderDao.deleteOrder(oid);
 	}
 }
